+++ tests/slurm_plugin/test_clustermgtd.py
@@ -594,20 +594,20 @@ def test_perform_health_check_actions(
     [
         (
             ClusterManager.HealthCheckTypes.scheduled_event,
-            [True, False, True, False],
-            [False, True, False, True],
+            [True, False, True, False, True, True, True],
+            [False, True, False, True, False, False, False],
             {"queue1-dy-c5xlarge-2", "queue1-st-c5xlarge-4"},
         ),
         (
             ClusterManager.HealthCheckTypes.ec2_health,
-            [True, False, True, False],
-            [False, True, True, False],
-            {"queue1-st-c5xlarge-1", "queue1-st-c5xlarge-3"},
+            [True, False, True, False, True, True, True],
+            [False, True, True, False, False, False, False],
+            {"queue1-st-c5xlarge-1", "queue1-st-c5xlarge-3", "queue1-st-c5xlarge-7"},
         ),
         (
             ClusterManager.HealthCheckTypes.ec2_health,
-            [False, False, False, False],
-            [False, True, False, True],
+            [False, False, False, False, False, False, False],
+            [False, True, False, True, False, True, True],
             {},
         ),
     ],
@@ -623,12 +623,24 @@ def test_handle_health_check(
     expected_failed_nodes,
     mocker,
 ):
-    # Define variable that will be used for all tests
+    # Define variables that will be used for all tests
+    mocker_current_time = datetime(2023, 1, 23, 18, 00, 0).astimezone(tz=timezone.utc)
     health_state_1 = EC2InstanceHealthState("id-1", "some_state", "some_status", "some_status", "some_event")
     health_state_2 = EC2InstanceHealthState("id-2", "some_state", "some_status", "some_status", "some_event")
     health_state_3 = EC2InstanceHealthState("id-3", "some_state", "some_status", "some_status", "some_event")
     health_state_4 = EC2InstanceHealthState("id-4", "some_state", "some_status", "some_status", "some_event")
-    placeholder_states = [health_state_1, health_state_2, health_state_3, health_state_4]
+    health_state_5 = EC2InstanceHealthState("id-5", "some_state", "some_status", "some_status", "some_event")
+    health_state_6 = EC2InstanceHealthState("id-6", "some_state", "some_status", "some_status", "some_event")
+    health_state_7 = EC2InstanceHealthState("id-7", "some_state", "some_status", "some_status", "some_event")
+    placeholder_states = [
+        health_state_1,
+        health_state_2,
+        health_state_3,
+        health_state_4,
+        health_state_5,
+        health_state_6,
+        health_state_7,
+    ]
     instance_id_to_active_node_map = {
         "id-1": StaticNode("queue1-st-c5xlarge-1", "ip-1", "host-1", "DOWN+CLOUD+NOT_RESPONDING", "queue1"),
         "id-2": DynamicNode(
@@ -636,6 +648,33 @@ def test_handle_health_check(
         ),
         "id-3": StaticNode("queue1-st-c5xlarge-3", "ip-3", "host-3", "ALLOCATED+CLOUD", "queue1"),
         "id-4": StaticNode("queue1-st-c5xlarge-4", "ip-4", "host-4", "ALLOCATED+CLOUD", "queue1"),
+        # This node is still rebooting, so it's not drained due to EC2 health checks failing
+        "id-5": StaticNode(
+            "queue1-st-c5xlarge-5",
+            "ip-5",
+            "host-5",
+            "DOWN+CLOUD+REBOOT_ISSUED",
+            "queue1",
+            slurmdstarttime=datetime(2023, 1, 23, 17, 56, 0).astimezone(tz=timezone.utc),
+        ),
+        # This node was rebooted very recently, so it's not drained due to EC2 health checks failing
+        "id-6": StaticNode(
+            "queue1-st-c5xlarge-6",
+            "ip-6",
+            "host-6",
+            "IDLE+CLOUD",
+            "queue1",
+            slurmdstarttime=datetime(2023, 1, 23, 17, 59, 0).astimezone(tz=timezone.utc),
+        ),
+        # This node was rebooted a long time ago, so it can be drained due to EC2 health checks failing
+        "id-7": StaticNode(
+            "queue1-st-c5xlarge-7",
+            "ip-7",
+            "host-7",
+            "IDLE+CLOUD",
+            "queue1",
+            slurmdstarttime=datetime(2023, 1, 23, 17, 56, 0).astimezone(tz=timezone.utc),
+        ),
     }
     mock_ec2_health_check = mocker.patch(
         "slurm_plugin.slurm_resources.EC2InstanceHealthState.fail_ec2_health_check",
@@ -647,11 +686,14 @@ def test_handle_health_check(
     )
     # Setup mocking
     mock_sync_config = SimpleNamespace(
-        health_check_timeout=10, protected_failure_count=2, insufficient_capacity_timeout=600
+        health_check_timeout=10,
+        health_check_timeout_after_slurmdstarttime=180,
+        protected_failure_count=2,
+        insufficient_capacity_timeout=600,
     )
 
     cluster_manager = ClusterManager(mock_sync_config)
-    cluster_manager._current_time = "some_current_time"
+    cluster_manager._current_time = mocker_current_time
     drain_node_mock = mocker.patch("slurm_plugin.clustermgtd.set_nodes_drain", autospec=True)
 
     # Run tests
@@ -662,8 +704,8 @@ def test_handle_health_check(
     else:
         mock_ec2_health_check.assert_has_calls(
             [
-                call("some_current_time", 10),
-                call("some_current_time", 10),
+                call(mocker_current_time, 10),
+                call(mocker_current_time, 10),
             ]
         ),
     if expected_failed_nodes:
@@ -710,7 +752,7 @@ def test_update_static_nodes_in_replacement(current_replacing_nodes, slurm_nodes
                     "hostname",
                     "IDLE+CLOUD",
                     "queue1",
-                    "(Code:ServiceUnavailable)Failure when resuming nodes",
+                    "(Code:ServiceUnavailable)Failure when resuming nodes [root@2023-01-31T21:24:55]",
                 ),
                 DynamicNode(
                     "queue1-dy-c5xlarge-2",
@@ -733,7 +775,7 @@ def test_update_static_nodes_in_replacement(current_replacing_nodes, slurm_nodes
                     "hostname",
                     "IDLE+CLOUD",
                     "queue1",
-                    "(Code:Exception)Failure when resuming nodes",
+                    "(Code:Exception)Failure when resuming nodes [root@2023-01-31T21:24:55]",
                 ),
                 DynamicNode(
                     "queue1-dy-c5xlarge-2",
@@ -741,7 +783,7 @@ def test_update_static_nodes_in_replacement(current_replacing_nodes, slurm_nodes
                     "hostname",
                     "IDLE+CLOUD",
                     "queue1",
-                    "(Code:InsufficientHostCapacity)Failure when resuming nodes",
+                    "(Code:InsufficientHostCapacity)Failure when resuming nodes [root@2023-01-31T21:24:55]",
                 ),
                 DynamicNode(
                     "queue1-dy-c5xlarge-3",
@@ -749,7 +791,7 @@ def test_update_static_nodes_in_replacement(current_replacing_nodes, slurm_nodes
                     "hostname",
                     "DOWN+CLOUD+NOT_RESPONDING+POWERING_UP",
                     "queue1",
-                    "(Code:InsufficientInstanceCapacity)Failure when resuming nodes",
+                    "(Code:InsufficientInstanceCapacity)Failure when resuming nodes [root@2023-01-31T21:24:55]",
                 ),
                 DynamicNode(
                     "queue1-dy-c5xlarge-4",
@@ -757,7 +799,7 @@ def test_update_static_nodes_in_replacement(current_replacing_nodes, slurm_nodes
                     "hostname",
                     "DOWN+CLOUD+NOT_RESPONDING+POWERING_UP",
                     "queue1",
-                    "(Code:InsufficientReservedInstanceCapacity)Failure when resuming nodes",
+                    "(Code:InsufficientReservedInstanceCapacity)Failure when resuming nodes [root@2023-01-31T21:24:55]",
                 ),
                 DynamicNode(
                     "queue1-dy-c5xlarge-5",
@@ -765,7 +807,7 @@ def test_update_static_nodes_in_replacement(current_replacing_nodes, slurm_nodes
                     "hostname",
                     "DOWN+CLOUD+POWERED_DOWN+NOT_RESPONDING",
                     "queue1",
-                    "(Code:MaxSpotInstanceCountExceeded)Failure when resuming nodes",
+                    "(Code:MaxSpotInstanceCountExceeded)Failure when resuming nodes [root@2023-01-31T21:24:55]",
                 ),
             ],
             [EC2Instance("id-1", "ip-1", "hostname", "some_launch_time"), None, None, None, None],
@@ -1107,7 +1149,7 @@ def test_handle_unhealthy_static_nodes(
                     "hostname",
                     "IDLE+CLOUD",
                     "queue",
-                    "(Code:MaxSpotInstanceCountExceeded)Failure when resuming nodes",
+                    "(Code:MaxSpotInstanceCountExceeded)Failure when resuming nodes [root@2023-01-31T21:24:55]",
                 ),  # ice dynamic node does not belong to unhealthy node when enable fast capacity failover
             ],
             [
@@ -1132,7 +1174,7 @@ def test_handle_unhealthy_static_nodes(
                             "hostname",
                             "IDLE+CLOUD",
                             "queue",
-                            "(Code:MaxSpotInstanceCountExceeded)Failure when resuming nodes",
+                            "(Code:MaxSpotInstanceCountExceeded)Failure when resuming nodes [root@2023-01-31T21:24:55]",
                         )
                     ]
                 }
@@ -1152,7 +1194,7 @@ def test_handle_unhealthy_static_nodes(
                     "hostname",
                     "DOWN+CLOUD+NOT_RESPONDING+POWERING_UP",
                     "queue",
-                    "(Code:MaxSpotInstanceCountExceeded)Failure when resuming nodes",
+                    "(Code:MaxSpotInstanceCountExceeded)Failure when resuming nodes [root@2023-01-31T21:24:55]",
                 ),  # unhealthy dynamic
                 StaticNode(
                     "queue2-st-c5xlarge-1", "ip-4", "hostname", "DOWN+CLOUD+NOT_RESPONDING", "queue1"
@@ -2623,7 +2665,8 @@ def test_find_bootstrap_failure_nodes(active_nodes, instances):
                             "nodehostname",
                             "COMPLETING+DRAIN",
                             "queue1",
-                            "(Code:InsufficientReservedInstanceCapacity)Failure when resuming nodes",
+                            "(Code:InsufficientReservedInstanceCapacity)Failure when resuming nodes "
+                            "[root@2023-01-31T21:24:55]",
                         ),  # unhealthy ice node
                         DynamicNode(
                             "queue1-dy-c5xlarge-3",
@@ -2631,7 +2674,7 @@ def test_find_bootstrap_failure_nodes(active_nodes, instances):
                             "nodehostname",
                             "DOWN+CLOUD",
                             "queue2",
-                            "(Code:InsufficientHostCapacity)Failure when resuming nodes",
+                            "(Code:InsufficientHostCapacity)Failure when resuming nodes [root@2023-01-31T21:24:55]",
                         ),  # unhealthy ice node
                     ],
                     "c4xlarge": [
@@ -2648,7 +2691,7 @@ def test_find_bootstrap_failure_nodes(active_nodes, instances):
                             "queue2-dy-c5large-1",
                             "IDLE+CLOUD+POWERED_DOWN",
                             "queue2",
-                            "(Code:InsufficientHostCapacity)Failure when resuming nodes",
+                            "(Code:InsufficientHostCapacity)Failure when resuming nodes [root@2023-01-31T21:24:55]",
                         ),  # powered down ice node
                         DynamicNode(
                             "queue2-dy-c5large-2",
@@ -2665,7 +2708,8 @@ def test_find_bootstrap_failure_nodes(active_nodes, instances):
                             "nodehostname",
                             "DOWN+CLOUD",
                             "queue2",
-                            "(Code:InsufficientHostCapacity)Temporarily disabling node due to insufficient capacity",
+                            "(Code:InsufficientHostCapacity)Temporarily disabling node due to insufficient capacity "
+                            "[root@2023-01-31T21:24:55]",
                         ),  # unhealthy ice node
                     ],
                 },
@@ -2679,7 +2723,8 @@ def test_find_bootstrap_failure_nodes(active_nodes, instances):
                             "nodehostname",
                             "COMPLETING+DRAIN",
                             "queue1",
-                            "(Code:InsufficientReservedInstanceCapacity)Failure when resuming nodes",
+                            "(Code:InsufficientReservedInstanceCapacity)Failure when resuming nodes "
+                            "[root@2023-01-31T21:24:55]",
                         ),  # unhealthy ice node
                         DynamicNode(
                             "queue1-dy-c5xlarge-3",
@@ -2687,7 +2732,7 @@ def test_find_bootstrap_failure_nodes(active_nodes, instances):
                             "nodehostname",
                             "DOWN+CLOUD",
                             "queue2",
-                            "(Code:InsufficientHostCapacity)Failure when resuming nodes",
+                            "(Code:InsufficientHostCapacity)Failure when resuming nodes [root@2023-01-31T21:24:55]",
                         ),  # unhealthy ice node
                     ],
                 },
@@ -2699,7 +2744,8 @@ def test_find_bootstrap_failure_nodes(active_nodes, instances):
                             "nodehostname",
                             "DOWN+CLOUD",
                             "queue2",
-                            "(Code:InsufficientHostCapacity)Temporarily disabling node due to insufficient capacity",
+                            "(Code:InsufficientHostCapacity)Temporarily disabling node due to insufficient capacity "
+                            "[root@2023-01-31T21:24:55]",
                         ),  # unhealthy ice node
                     ],
                 },
@@ -2781,7 +2827,8 @@ def test_handle_ice_nodes(
                             "nodehostname",
                             "COMPLETING+DRAIN",
                             "queue1",
-                            "(Code:InsufficientReservedInstanceCapacity)Failure when resuming nodes",
+                            "(Code:InsufficientReservedInstanceCapacity)Failure when resuming nodes "
+                            "[root@2023-01-31T21:24:55]",
                         ),
                     ],
                 },
@@ -2793,7 +2840,7 @@ def test_handle_ice_nodes(
                             "nodehostname",
                             "DOWN+CLOUD",
                             "queue2",
-                            "(Code:InsufficientHostCapacity)Failure when resuming nodes",
+                            "(Code:InsufficientHostCapacity)Failure when resuming nodes [root@2023-01-31T21:24:55]",
                         ),
                         DynamicNode(
                             "queue2-dy-c5large-2",
@@ -2801,7 +2848,8 @@ def test_handle_ice_nodes(
                             "nodehostname",
                             "DOWN+CLOUD",
                             "queue2",
-                            "(Code:InsufficientHostCapacity)Temporarily disabling node due to insufficient capacity",
+                            "(Code:InsufficientHostCapacity)Temporarily disabling node due to insufficient capacity "
+                            "[root@2023-01-31T21:24:55]",
                         ),
                     ],
                 },
@@ -2830,7 +2878,8 @@ def test_handle_ice_nodes(
                             "nodehostname",
                             "COMPLETING+DRAIN",
                             "queue1",
-                            "(Code:InsufficientReservedInstanceCapacity)Failure when resuming nodes",
+                            "(Code:InsufficientReservedInstanceCapacity)Failure when resuming nodes "
+                            "[root@2023-01-31T21:24:55]",
                         ),
                     ],
                 },
@@ -2842,7 +2891,7 @@ def test_handle_ice_nodes(
                             "nodehostname",
                             "DOWN+CLOUD",
                             "queue2",
-                            "(Code:InsufficientHostCapacity)Failure when resuming nodes",
+                            "(Code:InsufficientHostCapacity)Failure when resuming nodes [root@2023-01-31T21:24:55]",
                         ),
                         DynamicNode(
                             "queue2-dy-c5large-2",
@@ -2850,7 +2899,8 @@ def test_handle_ice_nodes(
                             "nodehostname",
                             "DOWN+CLOUD",
                             "queue2",
-                            "(Code:InsufficientHostCapacity)Temporarily disabling node due to insufficient capacity",
+                            "(Code:InsufficientHostCapacity)Temporarily disabling node due to insufficient capacity "
+                            "[root@2023-01-31T21:24:55]",
                         ),
                     ],
                 },
@@ -2931,7 +2981,8 @@ def test_update_insufficient_capacity_compute_resources(
                             "nodehostname",
                             "COMPLETING+DRAIN",
                             "queue1",
-                            "(Code:InsufficientReservedInstanceCapacity)Failure when resuming nodes",
+                            "(Code:InsufficientReservedInstanceCapacity)Failure when resuming nodes "
+                            "[root@2023-01-31T21:24:55]",
                         ),
                     ],
                 },
@@ -2943,7 +2994,7 @@ def test_update_insufficient_capacity_compute_resources(
                             "nodehostname",
                             "DOWN+CLOUD",
                             "queue2",
-                            "(Code:InsufficientHostCapacity)Failure when resuming nodes",
+                            "(Code:InsufficientHostCapacity)Failure when resuming nodes [root@2023-01-31T21:24:55]",
                         ),
                         DynamicNode(
                             "queue2-dy-c5large-2",
@@ -2951,7 +3002,8 @@ def test_update_insufficient_capacity_compute_resources(
                             "nodehostname",
                             "DOWN+CLOUD",
                             "queue2",
-                            "(Code:InsufficientHostCapacity)Temporarily disabling node due to insufficient capacity",
+                            "(Code:InsufficientHostCapacity)Temporarily disabling node due to insufficient capacity "
+                            "[root@2023-01-31T21:24:55]",
                         ),
                     ],
                 },
@@ -2992,7 +3044,8 @@ def test_update_insufficient_capacity_compute_resources(
                             "nodehostname",
                             "COMPLETING+DRAIN",
                             "queue1",
-                            "(Code:InsufficientReservedInstanceCapacity)Failure when resuming nodes",
+                            "(Code:InsufficientReservedInstanceCapacity)Failure when resuming nodes "
+                            "[root@2023-01-31T21:24:55]",
                         ),
                     ],
                 },
@@ -3004,7 +3057,7 @@ def test_update_insufficient_capacity_compute_resources(
                             "nodehostname",
                             "DOWN+CLOUD",
                             "queue2",
-                            "(Code:InsufficientHostCapacity)Failure when resuming nodes",
+                            "(Code:InsufficientHostCapacity)Failure when resuming nodes [root@2023-01-31T21:24:55]",
                         ),
                         DynamicNode(
                             "queue2-dy-c5large-2",
@@ -3012,7 +3065,8 @@ def test_update_insufficient_capacity_compute_resources(
                             "nodehostname",
                             "DOWN+CLOUD",
                             "queue2",
-                            "(Code:InsufficientHostCapacity)Temporarily disabling node due to insufficient capacity",
+                            "(Code:InsufficientHostCapacity)Temporarily disabling node due to insufficient capacity "
+                            "[root@2023-01-31T21:24:55]",
                         ),
                     ],
                 },
@@ -3024,7 +3078,7 @@ def test_update_insufficient_capacity_compute_resources(
                             "nodehostname",
                             "DOWN+CLOUD",
                             "queue2",
-                            "(Code:InsufficientHostCapacity)Failure when resuming nodes",
+                            "(Code:InsufficientHostCapacity)Failure when resuming nodes [root@2023-01-31T21:24:55]",
                         ),
                         DynamicNode(
                             "queue3-dy-c5large-2",
@@ -3032,7 +3086,8 @@ def test_update_insufficient_capacity_compute_resources(
                             "nodehostname",
                             "DOWN+CLOUD",
                             "queue2",
-                            "(Code:InsufficientHostCapacity)Temporarily disabling node due to insufficient capacity",
+                            "(Code:InsufficientHostCapacity)Temporarily disabling node due to insufficient capacity "
+                            "[root@2023-01-31T21:24:55]",
                         ),
                     ],
                 },
@@ -3132,7 +3187,7 @@ def test_reset_timeout_expired_compute_resources(
                             "queue2-dy-c5large-1",
                             "IDLE+CLOUD+POWERED_DOWN",
                             "queue2",
-                            "(Code:InsufficientHostCapacity)Failure when resuming nodes",
+                            "(Code:InsufficientHostCapacity)Failure when resuming nodes [root@2023-01-31T21:24:55]",
                         ),  # powered down ice node
                         DynamicNode(
                             "queue2-dy-c5large-2",
@@ -3236,7 +3291,7 @@ def test_set_ice_compute_resources_to_down(
                     "nodehostname",
                     "COMPLETING+DRAIN",
                     "queue1",
-                    "(Code:InsufficientReservedInstanceCapacity)Failure when resuming nodes",
+                    "(Code:InsufficientReservedInstanceCapacity)Failure when resuming nodes [root@2023-01-31T21:24:55]",
                 ),
                 DynamicNode(
                     "queue2-dy-c5large-1",
@@ -3244,7 +3299,7 @@ def test_set_ice_compute_resources_to_down(
                     "nodehostname",
                     "DOWN+CLOUD",
                     "queue2",
-                    "(Code:InsufficientHostCapacity)Failure when resuming nodes",
+                    "(Code:InsufficientHostCapacity)Failure when resuming nodes [root@2023-01-31T21:24:55]",
                 ),
                 DynamicNode(
                     "queue2-dy-c5large-2",
@@ -3252,7 +3307,8 @@ def test_set_ice_compute_resources_to_down(
                     "nodehostname",
                     "DOWN+CLOUD",
                     "queue2",
-                    "(Code:InsufficientHostCapacity)Temporarily disabling node due to insufficient capacity",
+                    "(Code:InsufficientHostCapacity)Temporarily disabling node due to insufficient capacity "
+                    "[root@2023-01-31T21:24:55]",
                 ),
             ],
             [
@@ -3280,7 +3336,8 @@ def test_set_ice_compute_resources_to_down(
                             "nodehostname",
                             "COMPLETING+DRAIN",
                             "queue1",
-                            "(Code:InsufficientReservedInstanceCapacity)Failure when resuming nodes",
+                            "(Code:InsufficientReservedInstanceCapacity)Failure when resuming nodes "
+                            "[root@2023-01-31T21:24:55]",
                         ),
                     ]
                 },
@@ -3292,7 +3349,7 @@ def test_set_ice_compute_resources_to_down(
                             "nodehostname",
                             "DOWN+CLOUD",
                             "queue2",
-                            "(Code:InsufficientHostCapacity)Failure when resuming nodes",
+                            "(Code:InsufficientHostCapacity)Failure when resuming nodes [root@2023-01-31T21:24:55]",
                         ),
                         DynamicNode(
                             "queue2-dy-c5large-2",
@@ -3300,7 +3357,8 @@ def test_set_ice_compute_resources_to_down(
                             "nodehostname",
                             "DOWN+CLOUD",
                             "queue2",
-                            "(Code:InsufficientHostCapacity)Temporarily disabling node due to insufficient capacity",
+                            "(Code:InsufficientHostCapacity)Temporarily disabling node due to insufficient capacity "
+                            "[root@2023-01-31T21:24:55]",
                         ),
                     ]
                 },
@@ -3345,7 +3403,7 @@ def test_set_ice_compute_resources_to_down(
                     "nodehostname",
                     "COMPLETING+DRAIN",
                     "queue1",
-                    "(Code:InsufficientReservedInstanceCapacity)Failure when resuming nodes",
+                    "(Code:InsufficientReservedInstanceCapacity)Failure when resuming nodes [root@2023-01-31T21:24:55]",
                 ),
                 DynamicNode(
                     "queue2-dy-c5large-1",
@@ -3353,7 +3411,7 @@ def test_set_ice_compute_resources_to_down(
                     "nodehostname",
                     "DOWN+CLOUD",
                     "queue2",
-                    "(Code:InsufficientHostCapacity)Failure when resuming nodes",
+                    "(Code:InsufficientHostCapacity)Failure when resuming nodes [root@2023-01-31T21:24:55]",
                 ),
                 DynamicNode(
                     "queue2-dy-c5large-2",
@@ -3361,7 +3419,8 @@ def test_set_ice_compute_resources_to_down(
                     "nodehostname",
                     "DOWN+CLOUD",
                     "queue2",
-                    "(Code:InsufficientHostCapacity)Temporarily disabling node due to insufficient capacity",
+                    "(Code:InsufficientHostCapacity)Temporarily disabling node due to insufficient capacity "
+                    "[root@2023-01-31T21:24:55]",
                 ),
             ],
             [
@@ -3379,7 +3438,7 @@ def test_set_ice_compute_resources_to_down(
                     "nodehostname",
                     "COMPLETING+DRAIN",
                     "queue1",
-                    "(Code:InsufficientReservedInstanceCapacity)Failure when resuming nodes",
+                    "(Code:InsufficientReservedInstanceCapacity)Failure when resuming nodes [root@2023-01-31T21:24:55]",
                 ),
                 DynamicNode(
                     "queue2-dy-c5large-1",
@@ -3387,7 +3446,7 @@ def test_set_ice_compute_resources_to_down(
                     "nodehostname",
                     "DOWN+CLOUD",
                     "queue2",
-                    "(Code:InsufficientHostCapacity)Failure when resuming nodes",
+                    "(Code:InsufficientHostCapacity)Failure when resuming nodes [root@2023-01-31T21:24:55]",
                 ),
                 DynamicNode(
                     "queue2-dy-c5large-2",
@@ -3395,7 +3454,8 @@ def test_set_ice_compute_resources_to_down(
                     "nodehostname",
                     "DOWN+CLOUD",
                     "queue2",
-                    "(Code:InsufficientHostCapacity)Temporarily disabling node due to insufficient capacity",
+                    "(Code:InsufficientHostCapacity)Temporarily disabling node due to insufficient capacity "
+                    "[root@2023-01-31T21:24:55]",
                 ),
             ],
             [