+++ CHANGELOG.md
@@ -9,6 +9,9 @@ This file is used to list changes made in each version of the aws-parallelcluste
 **ENHANCEMENTS**
 - Add logging of compute node console output to CloudWatch from head node on compute node bootstrap failure.
 - Add validators to prevent malicious string injection while calling the subprocess module.
+  
+- **BUG FIXES**
+- Fix an issue in clustermgtd that caused compute nodes rebooted via Slurm to be treated as unhealthy if the EC2 instance status checks fail (see https://github.com/aws/aws-parallelcluster/issues/4751)
 
 3.4.1
 ------